refactor: few methods
